[histogram-cuda] update the program and address the issue (https://github.com/ORNL/HeCBench/issues/267)

+++ src/histogram-cuda/histogram_compare.cu
@@ -220,7 +220,7 @@ void ReadTga(uchar4* &pixels, int &width, int &height, const char *filename)
     fseek(fptr, skip_bytes, SEEK_CUR);
 
     // Read the image
-    int pixel_bytes = header.bitsperpixel / 8;
+    size_t pixel_bytes = header.bitsperpixel / 8;
 
     // Allocate and initialize pixel data
     size_t image_bytes = width * height * sizeof(uchar4);
@@ -477,7 +477,7 @@ void TestMethods(
     {
         std::sort(timings.begin(), timings.end(), less_than_value());
         printf("Timings (us):\n");
-        for (int i = 0; i < timings.size(); i++)
+        for (size_t i = 0; i < timings.size(); i++)
         {
             double bandwidth = height * width * sizeof(PixelType) / timings[i].second / 1000;
             printf("\t %.3f %s (%.3f GB/s, %.3f%% peak)\n", timings[i].second, timings[i].first.c_str(), bandwidth, bandwidth / bandwidth_GBs * 100);
@@ -646,7 +646,7 @@ int main(int argc, char **argv)
         file_tests.push_back("austin");
         file_tests.push_back("cityscape");
 
-        for (int i = 0; i < file_tests.size(); ++i)
+        for (size_t i = 0; i < file_tests.size(); ++i)
         {
             printf("%s, ", file_tests[i].c_str());
             std::string filename = std::string("histogram/benchmark/") + file_tests[i] + ".tga";
